packages/dev/pyproject.toml
[project]
name = "dev"
version = "0.1.0"
description = "core 開発のためのパッケージ"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["api", "core", "pillow>=12.1.1"]

[dependency-groups]
dev = ["ruff>=0.15.6"]

[tool.uv.sources]
api = { workspace = true }
core = { workspace = true }
